Fix column unstacking bug in mesolve
There was a bug introduced by doing the column stacking in place; if
`e_ops` was given, but the `store_states` option was also set, then the
state would be unstacked, but the expectation values would expect it to
be unstacked.

We want to keep doing everything in-place rather than copying the data,
so we just insert some checks before we re-use the object.