Make nonmarkov use new data types
The change to pad_csr now makes it return a new CSR matrix, _not_ modify
the original in-place.  There's currently no support for modifying the
underlying buffers of the CSR type in-place anywhere in the data layer,
because we would have problems with numpy arrays that the user holds in
Python space that view onto the data.  This is unlikely to have any real
impact on performance since there's already not much concern for
avoiding copies in the rest of the solver.