Pass ownership to ndarray only if Dense is owner
`dense.wrap` can create a Dense object that does not own its own data,
but does not already have an ndarray backing.  This _should_ only be
used for quick wrapping of raw pointers (which largely shouldn't be used
at all, except in the data-layer transition) that should never reach
Python space.  If they do reach Python-space, there's likely a memory
leak,

Regardless, sometimes when debugging it's convenient to throw in a
`print(wrapped.as_ndarray())`, and if this is the case, there would be a
double-free error, unless we have this commit.