Add conversion to CSR from Dense and COO
The COO conversion is mostly just for use in pre-existing code - we use
it to effectively replace the old `COO_Matrix` struct and the
`COO_to_CSR` function.  It is not intended to be a full new data type on
the data layer.

Conversion from Dense in conjunction with dense.wrap or
dense.fast_from_numpy can replace the whole array of dense_*_to_csr_*
functions from before.