Convert brmesolve to new CSR type
The logic is completely unchanged within, but everything is swapped to
the new types.  The mathematical areas should hopefully be a little
clearer to read now, and there is only 2/3 of a small CSR increase in
memory usage within cop_super_term and cop_super_mult (less than +5%
within those functions, and has no bearing on speed); instead of getting
the conjugate by only conjugating the data array and feeding the parts
in separately we just use `c.conj()`.  This uses memcpy internally to
clone the structure arrays, so we pay very little penalty for it.

A couple of functions which were not used are removed.