Remove deprecated correlation API, use new data-layer
All deprecated functions have been removed for the new release, and
internally all the new data-layer types are now used when approriate.

The minor usage of eseries and ode2es is removed, but necessary
functionality is imported directly into private functions inside the
module, so that eseries and essolve can be removed.