Remove essolve and eseries
essolve in particular has been effectively unsupported since its
introduction, and has never even fulfilled large parts of its spec.

eseries was originally in place effectively to allow QuTiP to be a
drop-in replacement for the Quantum Optics toolbox in Matlab, but as the
library has evolved it has not been maintained, and almost all of its
functionality can now be replicated by QobjEvo, and much more besides.