Fix transpose bug in propagator
Before the data-layer switchover this part used `sp_reshape` and
transposed later, which got
switched to `unstack_columns` (which is equivalent to reshape followed
by transpose) but maintained the transpose.  This now removes the
"second" transpose, because it's a bug---it effectively makes the
time-evolution run backwards.

The row norming does not matter a great deal whether it's column-wise or
row-wise because the state is always Hermitian and the norm is the L2
norm (which is conserved by conjugation).