Update stochastic solvers for new data-types
In general we update them to use the "dispatching" forms of CQobjEvo.
Internally the stochastic solvers use memoryviews, so to avoid
duplicating mathematical code, we simply apply the `dense.wrap`
operation on all raw pointers before passing them into the dispatched
operations.

On my machine this has no speed penalty, and actually in certain cases
can result in up to a 2x speedup in some of the individual tests
(presumably because of more optimised mathematical operations).

Calling `dense.wrap` _does_ generate code which has to interact (very
briefly) with the Python interpreter, but does not allocate any memory.
If the Python interpreter interaction does at some point become a
problem, we could provide a `dense.rewrap` method which uses an existing
`Dense` instance, but resets all the internal elements to match the new
pointer structure.  This would avoid needing to hold the GIL after the
initial `Dense` instances have been created, and we would only ever need
two of these (though of course that has the potential to create a HUUUGE
race condition if we're not careful).