Cull unused code
All removed Cython is reimplemented in some way or another in the data
layer, and so is no longer needed separately.

cy.codegen and rhs_generate.rhs_generate have been unused since the
introduction of QobjEvo, and are consequently removed here.

The OpenMP components are not touched by this commit, and so currently
QuTiP will not build with `--with-openmp`.