Remove unused OpenMP components
This is not a full removal of all OpenMP code, and is not intended to
result in a complete removal of OpenMP from QuTiP 5, but we are
overhauling how OpenMP is handled within the library.

The files removed here were either already unused in the library (in
some parts due to the move to the data layer), or CQobjEvo OpenMP
variants which are all being consolidated properly into the data layer.

The OpenMP tests will still not compile and pass after this.