Remove Bloch-Redfield OpenMP code
This likely _will_ be permanent.  The only differences in the two
functions that br_omp.pyx exported are that they use the OpenMP version
of the CSR/Dense matrix multiplication.  In the new handling of OpenMP,
the data-layer will likely handle this on the developer's behalf, and so
there will be no need to import different functions to take advantage of
OpenMP.