Template CSR-vector matrix product over integers
Previously the file zspmv.cpp hard-coded the size of indexing integers
to be the system `int`, which meant that we always had to guarantee that
`idxint` was exactly equal to the system integer type, which rather
defeated the purpose of having it.

There were two possible ways of fixing this:
    1. template the function so all necessary specialisations are
       generated (though because Cython likes to link dynamically, we
       have to explicitly generate them)
    2. make the `idxint` type available in a C++ header file, so the
       files could include here
I went with method 1 because it was a little easier to get it to work,
and seemed less likely to have unpleasant interactions with numpy's C
API typedefs.  The cost is that we generate the same code three times
over, even though only one of them is actually used.  Since there are
very very few functions defined in C++ in QuTiP, this should be a fairly
trivial amount of storage, particularly as compared to the size of all
of the CPython extension modules we create.

I removed all the duplicate code intended to make things work between
MSVC and gcc/clang, where the only difference was how `__restrict` is
defined.  This is now just handled by a simple `#define` at the top, and
now the only code duplication is between manual vectorisation and not.