Add data-layer type inter-conversion
This object `data.to` will become the heart of all dispatch operations,
and really a large part of the driver of QuTiP data operations.  A type
is a "data-layer type" if it has conversions defined in this object, and
the dispatchers will use this to perform conversions when they do not
have a known specialisation for the input types.