Add multiple dispatch over types
This properly creates the multiple dispatch `Dispatcher` object.  It is
still quite rough around the edges, particularly in reporting failures,
but in "correct" uses it works.

An instance of `Dispatcher` is responsible for performing a single
mathematical function, for _all_ of the data types that the converter
`data.to` knows about.  It does this for all combinations of data types,
no matter how few specialisations the dispatcher actually knows about,
so that all of QuTiP can always use all possible data types, and the
developer or user adding a new type can do so at runtime, adding only
the one or two specialisations that they expect to use.

We use the converters defined in `data.to` to convert our inputs (and
possibly output) so that they match a specialisation we _do_ know about.
We build up a complete lookup table whenever we learn a new
specialisation or a new data type is added, so we don't have to do
anything at dispatch time.  Currently we do build this lookup table with
the brute-force algorithm of finding the lowest expected conversion
weight by just checking every legal input combination with every known
specialisation.  In the future, if many data types become first-class
QuTiP data types, this may become a speed and memory constraint.  If so,
we can switch to some smarter dynamic programming method, but until it's
proved to actually be a problem, brute-force much easier.