Set up dispatcher objects for data operations
This sets up (almost) all the dispatcher objects for the data-layer,
including writing all their docstrings and setting up the
specialisations.

The way the dispatchers are set up internally---manually defining the
signature and using _defer=True---involves much more boilerplate than
any end-user of QuTiP will need to write.

Inside the Cython extensions, we have to manually define the signatures
because `inspect` cannot extract the signature from a Cython object.
This is not the case for functions defined in Python-space.

Also, due to the interconnected nature of the type files (e.g. csr.pyx)
and the operation files (e.g.  add.pyx), the converter `data.to` cannot
be set-up with knowledge of the types before the operation modules are
imported, which leads to a problematic circular import.  We solve that
within the library initialisation by a special `_defer` keyword argument
in `add_specialisations`.  The user will not need to do this; they will
be able to add the new type to `data.to`, and then add specialisations
to the dispatchers they want without conflict, because `data.to` has
already been initialised.

The dispatchers for `eigs` and `ptrace` have not yet been implemented
because the way those two functions are written are not compatible with
the dispatcher model yet (but their interfaces will be changed to make
it work).