Remove public declaration in base.pxd
This stops the `base.h` file from being generated---the `public` keyword
was a mistake and shouldn't have been there.  In `pxd` files, `public`
makes the declaration public for other C interfaces (hence the creation
of the header file), whereas I had mistaken it for the use in `cdef
class` declarations where it makes the field public to Python.