Make COO to CSR conversion sum duplicates
This brings it inline with the scipy behaviour, and is generally a
convenience in other places as well.

We do not, however, achieve this in the same way scipy does; they first
directly convert a COO matrix into a CSR ignoring the duplicates, sort
the whole set, then another pass to sum the duplicates.  Instead, we use
the accumulation scatter/gather pattern on a row-by-row basis, with a
complete class to manage it, so it can be re-used in all other places
which construct a matrix row-by-row like `add_csr` and `matmul_csr`.