Add eigs dispatcher
Currently `eigs` doesn't fit the dispatcher form especially well; the
typical method of solving the eigenvalues should be to use the `Dense`
method _by default_, regardless of what the underlying type is.

We may want to change how this is accessed in the future; at the moment
we barely actually use the proper dispatcher, but since this is
effectively hard-coding the desired types, it doesn't really allow users
to specify their own data types which may have more efficient eigenvalue
methods.