Fix indexing error in csr.Accumulate.gather
Simple, silly error: using the "input" index on the "output" array.  We
would always write out the correct elements and there would always be
enough space to write them into, but we'd not write them into
consecutive addresses if any of them were zero.

This could then leave uninitialised memory which wasn't written to for
both the complex values and the column entries, which lead to segfaults
due to massive OOB access at the best of times, and silent incorrect
answers at the worst.