Add scale parameter to matmul dispatcher
To achieve this, we have to swap the order of `scale` and `out` in the
call structure, because the dispatchers need to have a "compatible"
signature, and they can't achieve this if `left, right, scale` are
passed in positional form.

This can possibly cause problems in Cython code; preliminary testing
seemed to suggest that Cython would do something wrong if I passed the
last parameter as a keyword (out) but not the penultimate one (scale),
which still had a default.  I was doing this in a period when there were
other problems, however, so this may not have been the exact cause.