Tidy up tidyup_csr
In previous versions of this function (before the fast_csr_matrix/CSR
switchover) the old C-struct CSR_Matrix would be resized if enough
elements had been deleted.  Since our CSR type may be backed by numpy
storage, this isn't as clearly a safe operation any more.  This may be a
solveable problem, but right now it's not a high priority as the
downsides to _not_ doing it are not large.