Add tidyup_dense
With QuTiP's default settings, this is one of the most important
functions to define to make sure that `Qobj` really uses a new data
type, since many mathematical operations will implicitly include a call
to `data.tidyup`.  Without a specialisation for it, this will end up in
the data getting converted to a different type immediately.