Use dispatchers across QuTiP
Mostly this is a simple change; we just remove the `_csr` or
`_csr_dense_dense` (or whatever) suffix from the data operations to make
them dispatching.

In a few places of old code (like in steadystate) we manually force the
data types into the CSR format so the rest of the code can continue
unimpeded.

In theory, this now means that QuTiP can use any data-layer type for
every single operation, and mix them completely freely.  In reality,
there may be a couple of places where CSR format is assumed in ways I
forgot to check for, and certainly the creation of objects still
_strongly_ favours creating CSR types.

The ergonomics of creating Qobjs and selecting dispatch methods (such as
for expm) where it may be beneficial to convert the input into another
type still need to be improved.