Remove sorted requirement from add_csr
Change the algorithm to use the scatter/gather pattern instead of the
previous sort-and-combine.  This does not change the resulting
complexity, because there is still a sorting step in the gather
operation, but it is more memory-efficient.

In practice, this is a ~2x speedup for reasonable matrices.  There is
maybe some other speedup available if the Accumulator object is
converted to pure C and all methods on it are inlined---the methods are
very lightweight, and should not generally need to be a separate call.