Make Accumulator a header-only include
This converts the CSR Accumulator pattern into a C struct, meaning
interactions with it no longer require holding the GIL.  All the
operations on Accumulator are moved in csr.pxd, meaning that they have
effectively become a header-only library, and now the functions can be
inlined properly between modules.

This transpired to not offer too much of a speed-up anywhere, but the C
struct approach is still the right way to approach it, because it allows
different operations to more easily operate on separate threads at the
same time (though of course the scatter/gather pattern itself is not
inherently thread-safe within the _same_ Accumulator).