Remove sort from Accumulator when scatters sorted
This removes an `O(n lg(n))` dependency from Accumulator when all the
calls to Accumulator.scatter that add a new entry are sorted.  This
means that scatters do not necessarily need to be exactly sorted, for
example the scatter pattern:
    0, 1, 2, 1, 1, 2, 0, 3
still counts as "sorted", because the only unsorted entries in the
pattern just accumulate onto previous values.

Algorithmically, this means that the sort can be removed from add_csr
when the left matrix is sorted and the right is "exclusively sorted"
with it (i.e. its non-shared values are sorted relative to each other).
This makes add_csr O(n) for sorted matrices, and O(n lg(n)) for unsorted
matrices, with the checks carried out during the scattering (no need for
a separate test).  To achieve this, we swap the add_csr algorithm back
to mutual iteration through the row, so that sorted matrices will always
be scattered in order, at the cost of sparse-matrix cache efficiency.