Fix buggy density test in permute.dimensions_csr
This was intended to estimate how many elements on average were in each
row, but the test wasn't exactly organised correctly and was subject to
integer overflow when dealing with large numbers of qubits (16 when
idxint is 32-bit).  It doesn't matter that the floating-point test will
lose a little bit of precision; we don't need to be 1ULP-accurate.