Fix permute.dimension_csr for sparse matrices
First remove the completely unnecessary triplicated calls to
`_Indexer.single()` by caching the results.  An otherwise sensible use
of the sparse code would get ruined by vastly over-using the call, which
is exactly what the special-case is trying to avoid.

Second, fix a buggy line: in `_dimensions_csr_sparse` there used to be
the lines (229--230) that's hard to spot in the diff because of the
release of the GIL in this commit:
    if n:
        out.row_index[index.single(row) + 1] = len
where the assignment should have been `n`, not `len`.  `len` is the
longest row seen before now, so the first row seen would also be zeroed
out by this function.  Also, a matrix with some rows having more
elements in that others would set the wrong values in later short rows,
probably causing a segfault.

Release the GIL when we don't need it for iteration - that's often a
free speedup in Cython code.

See #1382.