Remove use of div_t
This tended to cause annoying compilation problems on Windows in
particular; the compiler infrastructure there would typically not be
able to choose a suitable specialisation between div_t, ldiv_t and
lldiv_t when the base.idxint size was not the natural machine-int size.

This _could_ have been worked around with conditional compilation, but
the time saved by using div_t isn't worth it.