Fix _ptrace_dense when all subspaces are kept
For example, this fixes the case `A.ptrace(0)` where `A` is _not_ a
tensored Hilbert space; this has come up a few times in people's scripts
where they are looping over states that may or may not be tensored at a
given point.

This commit just fixes the underlying bug; `np.prod` called on an empty
Python list defaults to using a float, not an integer, so cannot then be
passed as part of an array's shape.