Make VERSION file for single source of truth
Now that wheels we be built from CI, and we will want to release direct
to PyPI/conda from there, it's more useful to have a single file that
can be simply read and parsed by the CI shell.  This also means that
release versioning changes need only affect a single file, and this file
can be populated from git tag or similar.

The "ISRELEASED" flag is now moved to either a commandline argument to
setup.py (--release), or by setting the environment variable
    CI_QUTIP_RELEASE
to be non-empty.  The environment variable strategy is better for CI
(hence the name), where tools like cibuildwheel can't pass on
command-line options to our setup.py.

We still create a git hash if not in release mode and git is available
(or add '+nogit' if not), in order to distinguish between release builds
and local versions in bugfixing.