Increase compatibility with PEP 517
Separate out most packaging data into the relevant new configuration
files, rather than mixing our data and code together in setup.py.  We
still rely on setuptools, and some parts of our Cython infrastructure
and versioning system still rely on us executing the setup.py file, but
with the additional data in pyproject.toml and setup.cfg, this can now
be handled in a more abstract manner by a range of build tools.

Cython is required at install time here, but we also make sure to
distribute the .cpp files with the source; in theory it should be
possible to build QuTiP without Cython as long as we do that, though
right now we're not set up for it.

We swap to package- and module-discovery techniques rather than
enumeration; it simplifies development of new packages a little, and
should be less error-prone.

You can now do
    pip install -e .
_or_
    python setup.py develop
in the git root to install QuTiP in editable mode.  The former will work
as long as a suitable Python and pip are available; all build
dependencies will be fulfilled during the run.

However,
    pip install -e .
will install build dependencies in a temporary venv, which typically
causes Cython to re-cythonise all .pyx and .pxd files, regardless of
whether they are out-of-date.  See, e.g.:
    https://discuss.python.org/t/pep-518-and-editable-mode-dont-install-already-satisfied-dependencies/3124/9
This may be improved in a future version of Cython.