Remove unused OpenMP code
Having these kicking around wasn't a problem before package discovery,
but with it getting turned on, it becomes a problem.  Also remove the
last remnants of OpenMP references: block_redfield.py imported some
functions but their actual calls were stubbed out, and the tests
imported the modules but unconditionally skipped them.  This caused
test errors, even though the code was never actually used.

OpenMP is still meant to be reinstated before QuTiP 5.0 is released, but
it won't be handled in the same manner.