Remove Qobj.__array__ for numpy 1.20 compatibility
Numpy 1.20 caused objects that defined __array__ to be coerced into
ndarray when calling something like
    np.array([qobj1, qobj2, qobj3], dtype=object)
The intent in this line (such as appears in Qobj.eigenstates and
qutrit_basis) is clearly to make a 1D array with 3 elements that are
Qobj.  Since the __array__ protocol is defined, what happens instead is
that a 3D array filled with complex numbers (but with dtype=object) is
created.  Lower versions of numpy did not perform coercion at this
level.

That particular example is arguable: it could be solved by returning a
regular list.  However, other libraries such as scipy's sparse matrices
also do not implement __array__, and having it in causes some very odd
behaviour.

Qobj.__array__ was originally added to make
    np.array(qobj)
and
    qobj.full()
equivalent.  This however also has the side-effect
of causing things like
    >>> np.sin(qutip.basis(2, 1))
    array([[0.        +0.j],
           [0.84147098+0.j]])
which is mostly nonsense.  The __array__ method (and later
__array_function__) are intended to give a class more control over how
numpy ufuncs and other numpy functions act on the class.  Qobj already
provides Qobj.full() to create a numpy array out of the matrix form of
the quantum object, and other forms such as np.exp(qobj1) are more
likely to be user error, intending something like qobj1.expm().

See:
    - gh-938, 48770ff
        original feature request and implementation.
    - gh-1433
        discussion of removal