Catch integer overflow segfault in zcsr_kron
The function _safe_multiply was meant to raise a Python exception if the
integer arithmetic overflowed.  However, neither it nor several of its
call sites had the full exception-handling mechanisms enabled.

In addition, extremely large tensor products such as
    tensor([basis(2, 0) for _ in [None]*31])
would segfault when using 32-bit integer arithmetic, because the size
calculation was not checked with _safe_multiply.