Re-use brtools.spec_func even in OpenMP mode
Typedef'ing two different versions of spec_func between the non- and
OpenMP version, even if they had identical prototypes,  was a bit
fragile.  Cython would mangle the two types differently (as they were in
different modules), and for some reason minor changes to other
prototypes in the module could cause type mismatches.

Importing the same typedef removes this possibility, and the type is
_meant_ to be the same whether or not we're in OpenMP mode anyway.