Remove unsafe dims, shape in Qobj.__init__
Having default keyword parameters be mutable is always a risk in Python,
since if they're modified then it changes the default.

Being able to detect the input as None also helps remove a couple of
places where ragged-list array-ification was occuring in numpy (it's
deprecated for a while now).  Since we can detect against the default
None with a simple `is` check, there's no need for implicit array
coercion with np.all or similar.