Fix mesolve segfault with bad initial state
If the initial state was found to be of type 'oper', no further testing
was done in initialisation of mesolve on the dimensions.  For example,
if a user incorrectly created a density matrix, such as by doing
    tensor(basis(20, 0), qeye(20)/20)
(presumably intending basis(20, 0).proj() or similar), it would have a
type of 'oper' inside Qobj, and mesolve would recognise it as a "proper"
density matrix, even though the shape was actually incorrect.  Such an
object would get passed down into the C layer, which does no shape
checking, without having its shape explicitly tested, and eventually
could cause a segfault in the matrix multiplication routine.

The correct check in this setup is to test the shapes when passed into
mesolve, if this is possible.  If the Hamiltonian/Liouvillian is
generated by a Python callable, we have to test every time the
time-dependence function is called.  A user could make a pathalogical
function that changes its dimensions on different times, and we cannot
risk a segfault.

Fix qutip/qutip#1456