Fail deployment build if confirmation incorrect
Previously the build would continue, just with a sentinel to show that
the deployment would not occur.  This was to allow the case when you
just wanted to build the wheels.

Now, you get the same behaviour if you don't enter anything.  If you
enter something, but it _doesn't_ match, then the build will fail
immediately.  This should just be a little easier, since you won't have
to manually cancel it if you get it wrong.

We have to have a slightly unusual dependency graph for our jobs
(everything depends on deploy_test) in order to fail as quickly as
possible.  As of 2021-03, GitHub Actions doesn't have a simpler method
of cancelling jobs that are executing on other runners, so this is the
cleanest solution without making external API calls.