Rewrite release guide to match new release process
The change to having GitHub Actions build all our wheels means that we
are now releasing rather more files than we used to, but all the
necessary files are built with no interaction from the administrator.

This updates the guide to list the new procedures.  It also adds in a
section about making a GitHub release, since that is convenient for
storing the large binary files that we link to.  This means we do not
have to add them to our git repository, which is really not designed to
handle this use-case.