Remove import version tests from __init__
These were needless busy work; Numpy and SciPy are mandatory
dependencies, and we have hard version requirements for them at install.
It's not our fault if stuff doesn't work because the user has
deliberately bypassed their package manager.

Also, this just meant one additional place to encode the version
requirements, rather than having a single source of truth.

We don't remove version tests for the _optional_ Cython, because the
user may just have an old version installed and not have realised it was
an optional dependency.