Remove unnecessary installation directory test
Release versions of QuTiP don't distribute setup.py, so this test could
only return True if you've loaded a dev version with the "ISRELEASED"
flag set to True.  The __file__ magic variable discovers where the QuTiP
installation path is, not where the user's current working directory is,
so the test does not appear to do what it was intended to.

Its current behaviour isn't necessarily an error in the current
distribution process - as of right now, the "release distribution"
documentation would actually have you _try_ to run QuTiP in this state
in order to run final tests before releasing.