Make qutip.__init__ less forgiving of exceptions
There were many cases in qutip.__init__ that ignored any exception
thrown, and a couple of places where `try` statements covered large
tracts of code that shouldn't have been allowed to fail.

For example, the only part of the Cython import process that is
acceptable to fail is the
    import Cython
statement (if the user doesn't have Cython, which is an optional
dependency).  If they _do_ have it, then any subsequent failure in
setting up the pyximport mechanism should be treated as an unhandled
exception.  From the user's side, they have fulfilled the requirement to
use Cython, so we shouldn't mask failures setting up the Cython
environment.  That makes it harder to debug issues for both us and the
users.

It is possible that this change will cause errors for previously working
code.  However, any issues turned up are more likely to be caused by
other factors, just we've been completely ignoring them up til now.