Move pyximport logic into one place
Previously the pyximport-specific logic was split between __init__ and
qutip.cy.pyxbuilder.  Since it's only relevant to pyximport, it should
all be in only one place, which makes the interplay bewteen different
elements easier to understand.

The CFLAGS check can be done on the first call to the pyximport
mechanism; we want to ensure that distutils is imported as late as
possible, because if anything wants to access setuptools before, we'll
get nasty warnings.  That's not as unlikely as it sounds, because Cython
may need to access it itself, and our new coefficients (coming in the
next major release) use setuptools as well.