Fix dnorm issues with sparse matrices (#1463)
Newer versions of cvxpy (1.1+) had introduced some problems for us with
sparse matrices: see cvxgrp/cvxpy#1159, which at the time of writing was
still unresolved.

This introduces a new method of sparse-matrix packing to workaround the
issue based on a method in the linked issue, and adds a dense-matrix
option to dnorm in cases where the sparse algebra does not work.  The
sparse form should typically be used, because the superoperator
representation of a quantum channel will usually cause the dimension to
be rather large, except when dealing with states of only one or two
qubits.

Also in cvxpy 1.1+, the '*' operator for matrix multiplication now
issues a DeprecationWarning in favour or using '@'.  cvxpy 1.0 also
supports '@', so the change is safe.

The memoisation was previously made over dnorm_problem, however really
the only slow operation was the call to tensor_swap.  The new method of
sparse-matrix packing makes the construction of the optimisation problem
strongly dependent on the sparsity structure of the supplied channel, so
memoisation over the full problem creation is not useful.  Instead, the
memoisation is moved to be only over the tensor_swap operation, which
also minimises the risk of modification of mutable values.

Co-authored-by: MrRobot2211 <felipebihaiek@gmail.com>
Co-authored-by: Jake Lishman <jake@binhbar.com>