Relax tolerance of test_diag_liou_mult
This complete test would have a failure about one run in every ten with
only one or two elements exceeding the tolerances on Macs when using
OpenBLAS.  These failures couldn't be reproduced when using MKL.  It is
very likely that this is related to us using `eig` instead of `eigh` in
this particular configuration (avoiding segfaults), which has a slight
effect on the tolerances, since `eig` is comparing complex tolerances
not real ones.

It should be safe to increase our tolerance marginally, because the
problem is not a logic failure in our code, but just a minor loss of
precision on one particular platform.

Fix gh-1431