Avoid matrix tidyup in trace norm
The first step of calculating the trace norm is to do
    A @ A.dag()
which we did internally in Qobj.norm.  This creates a temporary Qobj,
and those created from Qobj.__mul__ are subject to tidyup by default, so
this temporary structure may incorrectly get tidied to the zero matrix.

For example, the code
    qutip.qdiags(1e-7 * np.random.rand(10), 0).norm()
would previously give 0, clearly all elements of A @ A.dag() would be
less than 1e-14, and subject to tidyup.

We avoid the call that may be tidied up by jumping straight to CSR-CSR
matrix multiplication, and set isherm=True without going via Qobj, since
the operation is Hermitian for any matrix.

Fix gh-952