Refactor dnorm tests and allow reruns
dnorm tests have historically been very flaky, and for as long as we're
using random tests (which we possibly shouldn't be), we always run the
risk of temperamental failures.  In particular, the semidefinite solvers
used in dnorm fail every now and again on random states; we historically
haven't treated this as a bug, since it's pretty similar to an
integration failing to converge because the system was too stiff.  It
means that the user has to adjust some settings and try again.

To prevent this sort of error from failing our test suite, however, we
can mark the tests as being allowed to rerun twice on a failure.  We had
a relatively small sample to judge the test failure rate by, but my very
very approximate guess is that we'd have a failure about 1 in 10 runs on
Travis.  We counter this with two strategies:
    1. parametrise the tests in pytest style, so each random repetition
       is a separate instance, and known to pytest (some of this was
       done in earlier commits)
    2. allow 2 reruns of every dnorm test
The two of these together should remove all random failures; since all
random loops have been moved into pytest-handled repeats, there are far
fewer calls to dnorm handled within each test. This means less chance
that an individual test run will fail, but a slightly higher chance the
entire suite will.  Now with the test re-running a single failure point
is tried again up to twice more, to determine if it was just a one-off,
or if the test is truly broken.

The reruns are handled by a pytest plugin pytest-rerunfailures.  It's
quite nontrivial behaviour to hook this in to the pytest mechanism, so
better to use an external dependency to do it than to vendor in a custom
version.