Fix outdated star-imports from scipy and pylab
Things like scipy.sin have worked for quite a while, but are remnants of
the old Matlab-like flat-namespace approach, where people just
star-imported everything with no regard to name collisions or anything
like this.  pylab is an example of this; the correct modern approach is
to use qualified access through matplotlib.pyplot.

At any rate, scipy.sin and friends issues tonnes of deprecation
warnings, so we remove their use.