Fix tic location warnings in vizualisation.py
These were warnings raised with no user fault when calling hinton and a
couple of other cases, due to mixing a variable tic locator with a fixed
label creator.  We know ahead of time where our tics should be, so it's
simple to swap it over to a fixed locator.