Fix docstrings to be valid numpydoc and rst
In various places over time, we have accumulated docstrings that do not
fully adhere to the numpydoc and Sphinx' reStructuredText conventions,
resulting in many many warnings being generated when building the
documentation.

This (rather large) commit fixes all current warnings in building the
HTML and LaTeX documentation.  This involved several different
components:

- removing unnecessary "methods" sections from class docstrings (the
  correct location for numpydoc is just to document each method
  separately).

- removing the use of the `|` character, which Sphinx interprets as part
  of a text-replacement macro.  Since our documentation is predominantly
  made to be read online and these almost invariably appear as
  text-based braket notation, we can swap them over to LaTeX blocks with
  the :math: directive.

- converting places where "meaningful space" was used to logically
  separate objects into proper structures like lists.

This wasn't intended to completely rewrite any docstrings, or actually
_change_ any documentation, nor has it necessarily fixed all formatting
problems in API documentation.  This was a decent first stab at it
though, and fixes all the actual warnings.