Add loose files into built guide
This includes building in a few help files that had previously been
just loose in the repository.  We put them in a separate admin section.

We also make sure to exclude source files for the gallery extension, so
sphinx doesn't see the un-processed rst files and complain about them
not being linked.