Rewrite tree plot to have complete module discovery
It now recurses fully through the module and access all names from
within Python, so we get much easier and more accurate access to all the
objects by introspection.

This also changes the tree to have more than one layer, so we have to
change the styles a little to accommodate that.