Relax eigenvalue sum tolerance
This test is meant to be a sanity test on the eigenvalues that come out,
without actually having the true eigenvalues to compare against.  Since
the sum is subject to potential compounding of errors (although we
expect them to cancel on average), the tolerance should represent the
total uncertainty in the answer, not just a single eigenvalue.