Workaround scipy<1.5 / MKL bug in eigh (#1498)
SciPy used to make invalid calls to underlying MKL functions in versions
before 1.5 when calling eigh on large (dimension >= ~2000) matrices.  We
can't drop support for version 1.4 yet because at the time of writing
it's well under 2 years old, and we don't want to limit versioning too
much.

The additional dependence on packaging is the "proper" way of comparing
version information.

The additional test was sufficient to reproduce the issues in a minimal
amount of time.  This test is meant to be a sanity test on the
eigenvalues that come out, without actually having the true eigenvalues
to compare against.  Since the sum is subject to potential compounding
of errors (although we expect them to cancel on average), the tolerance
should represent the total uncertainty in the answer, not just a single
eigenvalue.

Co-authored-by: Jake Lishman <jake@binhbar.com>