Filter DeprecationWarnings inside correlation
The eseries-based solvers for correlation and spectrum are not being
removed in QuTiP 5.0, even though the public-facing eseries and ode2es
_are_ being removed.  This means that calling correlation_2op_1t and
spectrum should not generate warnings to the user, no matter what their
warning settings were.

The warnings would typically have been filtered when running in the
Python interpreter or Jupyter notebooks.  They did, however, appear when
running the tests, or building the documentation, making it harder to
see what else was going on.