Use linalg.eigvalsh instead of sparse.linalg.eigsh
For one, the passed object was actually a numpy array, not a sparse
matrix.  Two, it appears that _all_ eigenvalues were needed; sparse
routines typically cannot produce all eigenvalues, and certainly not to
any great tolerance.  Instead, we can use a general-purpose,
eigenvalue-only dense routine, since we already know the matrix must be
able to fit in memory.  This prevents a RuntimeWarning appearing from
ARPACK.